bleem-ai-api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
rembg
onnxruntime
aiosqlite
google-generativeai
pydantic
httpx
python-dotenv
aiomysql
PyMySQL
PyJWT
cryptography